# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmniReferral API",
        "description": "REST API for the OmniReferral real estate referral and lead management platform.\n\nProvides authentication, user management, realtor profiles, GoHighLevel CRM integration,\nonboarding webhooks, and administrative functions.\n\n## Authentication\nMost endpoints require a Bearer token obtained via the Login or Register endpoints.\nPass the token in the `Authorization` header as `Bearer <token>`.\n\n## Webhooks\nGoHighLevel webhook endpoints are public (CSRF exempt) and authenticated via\nthe `X-OmniReferral-Webhook` header secret.",
        "contact": {
            "email": "admin@omnireferrals.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://omnireferrals.com",
            "description": "OmniReferral Production API Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/admin/mail-settings": {
            "get": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "View mail settings",
                "description": "Shows current effective mail configuration (DB + .env fallback). Admin access required.",
                "operationId": "ed0104dcf89bf407b3095a4124a8c150",
                "responses": {
                    "200": {
                        "description": "Mail settings page"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "Update mail configuration",
                "description": "Updates SMTP/mail driver credentials. Super admin only. Password is stored encrypted.",
                "operationId": "3ac24a40cf66490bd4e0028b4627f1c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MailSettings"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect with success message"
                    },
                    "403": {
                        "description": "Forbidden - super admin only"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/mail-settings/test-connection": {
            "post": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "Test SMTP connection",
                "description": "Attempts to connect to the configured SMTP server. Returns success or error message.",
                "operationId": "b3601ca3399297c9f2fab2aeeab29e17",
                "responses": {
                    "200": {
                        "description": "Connection test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/mail-settings/test": {
            "post": {
                "tags": [
                    "Admin",
                    "Email System"
                ],
                "summary": "Send test email",
                "description": "Sends a raw test email to verify mail delivery is working end-to-end.",
                "operationId": "3bc66f77632fa0e7d71cc3e7813ba41b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MailSettings": {
                "properties": {
                    "mailer": {
                        "type": "string",
                        "enum": [
                            "smtp",
                            "sendmail",
                            "ses",
                            "postmark",
                            "resend",
                            "log",
                            "mailgun"
                        ]
                    },
                    "host": {
                        "type": "string"
                    },
                    "port": {
                        "type": "integer"
                    },
                    "encryption": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "description": "Leave blank to keep existing",
                        "type": "string",
                        "format": "password"
                    },
                    "from_address": {
                        "type": "string",
                        "format": "email"
                    },
                    "from_name": {
                        "type": "string"
                    },
                    "credentials_from_address": {
                        "type": "string",
                        "format": "email"
                    },
                    "credentials_from_name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "sanctum",
                "description": "Enter Sanctum token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}